requirements.txt
streamlit==1.28.1
deepgram-sdk==3.5.0
python-dotenv==1.0.0
